+++ buildSrc/src/main/java/BuildConfig.kt
@@ -1,0 +1,7 @@
+
+object buildConfig {
+    val compileSdk = 31
+    val minSdk = 15
+    val targetSdk = compileSdk
+    val libraryVersion = "3.5.0"
+}